requirements.txt
Django==4.1.7
djangorestframework==3.14.0
environs==9.5.0
django-filter==22.1
